Adding NLA exercise for Best Rank K approximation of a Matrix.
This is an exercise utilizing the properties of the SVD factorization of a matrix in order to approximate an image. In this case, the image's pixels can be thought of as a matrix, and we see that we can approximate a pretty good concept of an original image with a relatively low rank matrix.